requirements.txt
Flask==3.0.2
pymongo==4.6.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
pyjwt==2.6.0
redis==5.0.1
celery==4.4.0

